Cambio de las relaciones de los filtros por left join.

+++ src/main/java/com/example/demo/DTOs/FiltrosProductoDTO/FiltroProducto.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class FiltroProducto {
     public String nombreCategoria;
     public Integer porcentajeDescuento;
-    public Double precioMin;
-    public Double precioMax;
+    public BigDecimal precioMin;
+    public BigDecimal precioMax;
 }
